Fix clang warning: implicit conversion from 'long' to 'double'
changes value from 9223372036854775807 to 9223372036854775808
[-Wimplicit-const-int-float-conversion]
        if(bytes < 0 || bytes > INT64_MAX)
                              ~ ^~~~~~~~~
/usr/include/x86/_stdint.h:90:19: note: expanded from macro 'INT64_MAX'
 #define INT64_MAX       0x7fffffffffffffff
                         ^~~~~~~~~~~~~~~~~~

+++ src/xpdev/genwrap.c
@@ -281,7 +281,7 @@ int64_t parse_byte_count(const char* str, ulong unit)
 	}
 	if(unit > 1)
 		bytes /= unit;
-	if(bytes < 0 || bytes > INT64_MAX)
+	if(bytes < 0 || bytes > (double)INT64_MAX)
 		return INT64_MAX;
 	return (int64_t)bytes;
 }
